// src/auth/enums/user.enums.ts
export enum UserRole {
  SALES_REPRESENTATIVE = 'Sales Representative',
  SALES_MANAGER = 'Sales Manager',
  MANAGEMENT = 'Management',
  OTHER = 'other',
}

export enum CompanySize {
  SIZE_1_10 = '1-10',
  SIZE_11_50 = '11-50',
  SIZE_51_200 = '51-200',
  SIZE_201_500 = '201-500',
  SIZE_501_1000 = '501-1000',
  SIZE_1000_PLUS = '1000+',
}

export enum Industry {
  REAL_ESTATE = 'real-estate',
  AUTOMOTIVE = 'automotive',
  TEXTILE = 'textile',
  ELECTRONICS = 'electronics',
  PHARMACEUTICALS = 'pharmaceuticals',
  FOOD_BEVERAGE = 'food-beverage',
  CONSTRUCTION = 'construction',
  FURNITURE = 'furniture',
  AGRICULTURE = 'agriculture',
  CHEMICALS = 'chemicals',
  AEROSPACE = 'aerospace',
  HEALTHCARE = 'healthcare',
  RETAIL = 'retail',
  ENERGY_OIL_GAS = 'energy-oil-gas',
  RENEWABLE_ENERGY = 'renewable-energy',
  PLASTICS = 'plastics',
  PAPER_PULP = 'paper-pulp',
  TELECOMMUNICATIONS = 'telecommunications',
  MINING = 'mining',
  MARINE = 'marine',
  JEWELRY = 'jewelry',
  PRINTING_PUBLISHING = 'printing-publishing',
  COSMETICS = 'cosmetics',
  LOGISTICS = 'logistics',
  EDUCATION = 'education',
  MANPOWER_SERVICES = 'manpower-services',
  OTHER = 'other',
}